Implement GyroTest and placeholder for Submerged2025
Each of the source files in here can be copied to main.py and then compiled and run.  The main.py file is never to have the ultimate source code when it is pushed back to the repository.  Means any changes made in main.py should be copied back to the original source file before pushing.